convert labels to long

+++ SegFormer/engine.py
@@ -28,7 +28,7 @@ def train_one_epoch(args, model, optimizer, loss_fn, dataloader, scheduler,
                 loss = loss_fn(logits, lbl)
         else:
             logits = model(img)
-            lbl = lbl.squeeze(1)
+            lbl = lbl.squeeze(1).long()
             print(lbl)
             print(logits)
             loss = loss_fn(logits, lbl)
